translate Chinese to English

+++ dev/metrics/confusion_matrix.py
@@ -16,22 +16,22 @@ def confusion_matrix(
     :param normalize: {'true', 'pred', 'all', None}.
     :return:
     """
-    dtype = torch.float32
+    dtype = torch.long
     y_true = torch.as_tensor(y_true, dtype=dtype)
     y_pred = torch.as_tensor(y_pred, dtype=dtype)
     #
     n_classes = torch.max(y_true).item() + 1
-    # 计算cm
-    yt_yp_idxs = y_true * n_classes + y_pred  # 混淆矩阵.flat的索引
+    # calculate cm
+    yt_yp_idxs = y_true * n_classes + y_pred  # indices of cm.flat. cm[y_true, y_pred]
     cm = torch.bincount(yt_yp_idxs, minlength=n_classes * n_classes)
     cm = cm.reshape(n_classes, n_classes)
-    # 处理normalize. None 不做处理
-    if normalize == "true":  # 每个true标签(row)和=1
+    # for normalize. None does no processing
+    if normalize == "true":  # The sum of each true tag (row) =1
         cm = cm / torch.sum(cm, dim=1, keepdim=True)
     elif normalize == "pred":
         cm = cm / torch.sum(cm, dim=0, keepdim=True)
     elif normalize == "all":
         cm = cm / cm.sum()
-    # 处理除0导致的nan
+    # Handles nans caused by division by zero
     cm = torch.nan_to_num(cm)
     return cm
